Fix miter artifacts and add perspective option.
It can still look a bit flakey for very thick lines, but no more infinite spikes. I am not exactly sure what unit thickness is in when perspective is enabled and that the thickness does not depend on screen size. In any case, there should probably be a different shader for true world space lines. This should probably only be used to add a bit of perspective effect to screen space lines.